Multi-Modal-RAG/requirements.txt
fastapi
markdown
python-dotenv
pypdf
langchain
langchain-core
langchain-community
langchain-huggingface
langchain_text_splitters
langchain_google_genai
pinecone
pinecone-text
uvicorn[standard]
jinja2
python-multipart
pymongo
redis
Pillow
pymupdf
transformers
numpy
aiohttp
cachetools
